requirements.txt
tiktoken==0.7.0
langchain-core==0.2.9
langchain-cli==0.0.25
langchain-openai==0.1.8
langchain-experimental==0.0.61
langserve==0.2.2

langfuse==2.36.1
sentry-sdk==2.6.0
python-dotenv==1.0.1
pydantic==1.10.13
uvicorn==0.23.2
bs4==0.0.2
## The following requirements were added by pip freeze:
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bracex==2.4
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.110.3
frozenlist==1.4.1
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
griffe==0.47.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.5
langchain-community==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.80
libcst==1.4.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-autorefs==1.0.1
mkdocs-get-deps==0.2.0
mkdocs-include-markdown-plugin==6.2.0
mkdocstrings==0.25.1
mkdocstrings-python==1.10.5
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.0
orjson==3.10.5
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
Pygments==2.18.0
pymdown-extensions==10.8.1
pyproject-toml==0.0.10
pytest==8.2.2
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.9
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.4.1
toml==0.10.2
tomlkit==0.12.5
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
wcmatch==8.5.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
